requirements.txt
streamlit>=1.39.0
streamlit-chat
boto3
asyncio
strands-agents
strands-agents-tools
mcp
langchain_experimental
packaging>=16.8,<24
protobuf>=3.20,<5
python-dotenv
requests
fastapi
uvicorn
pydantic
reportlab